Fix typo

+++ Makefile
@@ -412,13 +412,13 @@ nodenv: ## Install nodenv node-build
 wordpress: ## Deploy wordpress
 	mkdir -p ${HOME}/src/github.com/masasam
 	sudo pacman -S nginx php-fpm php-gd
+	sudo ln -vsf ${PWD}/etc/php/php-fpm.d/www.conf   /etc/php/php-fpm.d/www.conf
 	wget https://ja.wordpress.org/wordpress-latest-ja.tar.gz
 	tar zxvf wordpress-latest-ja.tar.gz
-	rm wordpress-latest-ja.tar.gz
-	mv wordpress ${HOME}/src/github.com/masasam/
+	test -f	wordpress-latest-ja.tar.gz && rm wordpress-latest-ja.tar.gz
+	test -d ${HOME}/src/github.com/masasam/wordpress/wp-admin || mv wordpress ${HOME}/src/github.com/masasam/
 	echo 'create database wp' | mysql -u root
 	echo "grant all privileges on wp.* to wp@localhost identified by 'password';FLUSH PRIVILEGES;" | mysql -u root
-	sudo ln -vsf ${PWD}/etc/php/php-fpm.d/www.conf   /etc/php/php-fpm.d/www.conf
 
 emacs-devel: # Install development version of emacs
 	cd ${HOME}/src/github.com/masasam;\
